requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
build==0.8.0
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
cryptography==36.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fonttools==4.30.0
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numexpr==2.8.1
numpy==1.21.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.15.8
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2020.1
pyu2f==0.1.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
samplerate==0.1.0
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==59.5.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.6.0
tomli==2.0.1
torch>=1.11.0
torchaudio>=0.11.0
torchmetrics~=0.9.3
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typed-ast==1.5.2
typing-extensions==3.10.0.2
unicodedata2==14.0.0
urllib3==1.26.8
wandb>=0.12.20
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0

plotly~=5.9.0